+++ packages/webview_flutter/webview_flutter_platform_interface/test/src/method_channel/webview_method_channel_test.dart
@@ -13,7 +13,9 @@ import 'package:webview_flutter_platform_interface/webview_flutter_platform_inte
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
-  group('Tests on `plugin.flutter.io/webview_<channel_id>` channel', () {
+  group(
+      'Tests on `plugin.flutter.io/webview_<channel_id>` channel dart->native',
+      () {
     const int channelId = 1;
     const MethodChannel channel =
         MethodChannel('plugins.flutter.io/webview_$channelId');
@@ -55,6 +57,61 @@ void main() {
 
     tearDown(() {
       log.clear();
+    });
+
+    test('loadFile', () async {
+      await webViewPlatform.loadFile(
+        '/folder/asset.html',
+      );
+
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall(
+            'loadFile',
+            arguments: '/folder/asset.html',
+          ),
+        ],
+      );
+    });
+
+    test('loadHtmlString without base URL', () async {
+      await webViewPlatform.loadHtmlString(
+        'Test HTML string',
+      );
+
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall(
+            'loadHtmlString',
+            arguments: <String, String?>{
+              'html': 'Test HTML string',
+              'baseUrl': null,
+            },
+          ),
+        ],
+      );
+    });
+
+    test('loadHtmlString without base URL', () async {
+      await webViewPlatform.loadHtmlString(
+        'Test HTML string',
+        baseUrl: 'https://flutter.dev',
+      );
+
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall(
+            'loadHtmlString',
+            arguments: <String, String?>{
+              'html': 'Test HTML string',
+              'baseUrl': 'https://flutter.dev',
+            },
+          ),
+        ],
+      );
     });
 
     test('loadUrl with headers', () async {
@@ -496,6 +553,40 @@ void main() {
           ),
         ],
       );
+    });
+  });
+
+  group(
+      'Tests on `plugin.flutter.io/webview_<channel_id>` channel native->dart',
+      () {
+    const int channelId = 1;
+    final WebViewPlatformCallbacksHandler callbacksHandler =
+        MockWebViewPlatformCallbacksHandler();
+    final JavascriptChannelRegistry javascriptChannelRegistry =
+        MockJavascriptChannelRegistry();
+
+    MethodChannelWebViewPlatform(
+      channelId,
+      callbacksHandler,
+      javascriptChannelRegistry,
+    );
+
+    tearDown(() {
+      reset(callbacksHandler);
+    });
+
+    test('onUrlChanged', () async {
+      // Run
+      await ServicesBinding.instance!.defaultBinaryMessenger
+          .handlePlatformMessage(
+        'plugins.flutter.io/webview_$channelId',
+        StandardMethodCodec()
+            .encodeMethodCall(MethodCall('onUrlChanged', {'url': 'Test Url'})),
+        null,
+      );
+
+      // Verify
+      verify(callbacksHandler.onUrlChanged('Test Url'));
     });
   });